+++ changelog.md
@@ -1,9 +1,9 @@
 # Changelog
 
 ### Latest
-- Add support for different labels for each track-like group [!688](https://gitlab.cern.ch/atlas-flavor-tagging-tools/algorithms/umami/-/merge_requests/688)
 
-- 
+- Make track weight variable customizable [!712](https://gitlab.cern.ch/atlas-flavor-tagging-tools/algorithms/umami/-/merge_requests/712)
+- Add support for different labels for each track-like group [!688](https://gitlab.cern.ch/atlas-flavor-tagging-tools/algorithms/umami/-/merge_requests/688)
 
 ### [v0.19](https://gitlab.cern.ch/atlas-flavor-tagging-tools/algorithms/umami/-/tags/0.19) (20.03.2023)